Fix currency_value_class_info error
Fixes #66

Originally the method was passed as a string to the if condition.
This is equivalent to passing the method itself rather than using
the result of the method call in the if condition.

When adding support for Rails 5, the string was replaced with the
method call which caused the bug. This has been replace with a lambda
which represents the method itself rather than the result of the method
being called.